+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PATCH` verb
 - Overloading to support a message body
 
+### Changed
+
+- Error handling improvements everywhere.
+
 ## [06-Mar-2021] v0.1.5
 
 ### Added